requirements.txt
python-telegram-bot==13.13
python-telegram-bot[job-queue]==13.13
python-dateutil==2.8.2
Pillow==9.5.0
opencv-python==4.7.0.72
pdfminer.six==20221105
PyPDF2==3.0.0
numpy==1.23.5
python-dotenv==1.0.0
Werkzeug==2.3.0
requests==2.31.0
matplotlib==3.7.1
pandas==2.0.0
openpyxl==3.1.2
SpeechRecognition==3.10.0
pydub==0.25.1
